test: update pin modal spec
